react/post class/rc09-10_useEffect-axios-example/server/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
